claude-3.5-api-tutorial-main/requirements.txt
anthropic==0.30.0
pytest==8.2.2
python-dotenv==1.0.1